rename .ds to use on client and server
